requirements.txt
streamlit
pandas
Pillow


